+++ emcee/tests.py
@@ -83,6 +83,8 @@ def ln_flat(x):
 class Tests:
 
     def setUp(self):
+        np.random.seed(42)
+
         self.nwalkers = 100
         self.ndim = 5
 
@@ -149,7 +151,7 @@ class Tests:
             + 0.5 * np.log(np.linalg.det(self.cov))
 
         lnZ, dlnZ = self.sampler.thermodynamic_integration_log_evidence()
-        print(self.sampler.get_autocorr_time())
+        print(self.sampler.get_autocorr_time(c=2))
 
         assert np.abs(lnZ - (gaussian_integral - log_volume)) < 3 * dlnZ, \
             ("evidence incorrect: {0:g} versus correct {1:g} (uncertainty "
@@ -163,6 +165,13 @@ class Tests:
         self.sampler = MHSampler(self.cov, self.ndim, lnprob_gaussian,
                                  args=[self.icov])
         self.check_sampler(N=self.N * self.nwalkers, p0=self.p0[0])
+
+    def test_mh_unif(self):
+        f = lambda x: 0.0 if np.all((0.0 <= x) & (x <= 1.0)) else -np.inf
+        self.sampler = MHSampler(self.cov, self.ndim, f)
+        self.sampler.run_mcmc(np.random.rand(self.ndim), 100)
+        chain = self.sampler.chain
+        assert np.any(np.abs(np.diff(chain, axis=0)) > 0.0)
 
     def test_ensemble(self):
         self.sampler = EnsembleSampler(self.nwalkers, self.ndim,
@@ -230,11 +239,11 @@ class Tests:
         else:
             assert False, "The sampler should have failed by now."
 
-    # def test_parallel(self):
-    #     self.sampler = EnsembleSampler(self.nwalkers, self.ndim,
-    #                                    lnprob_gaussian, args=[self.icov],
-    #                                    threads=2)
-    #     self.check_sampler()
+    def test_parallel(self):
+        self.sampler = EnsembleSampler(self.nwalkers, self.ndim,
+                                       lnprob_gaussian, args=[self.icov],
+                                       threads=2)
+        self.check_sampler()
 
     def test_pt_sampler(self):
         cutoff = 10.0
@@ -259,3 +268,23 @@ class Tests:
         # Make sure that the blobs aren't all the same.
         blobs = self.sampler.blobs
         assert np.any([blobs[-1] != blobs[i] for i in range(len(blobs) - 1)])
+
+    def test_run_mcmc_resume(self):
+
+        self.sampler = s = EnsembleSampler(self.nwalkers, self.ndim,
+                                           lnprob_gaussian, args=[self.icov])
+
+        # first time around need to specify p0
+        try:
+            s.run_mcmc(None, self.N)
+        except ValueError:
+            pass
+
+        s.run_mcmc(self.p0, N=self.N)
+        assert s.chain.shape[1] == self.N
+
+        # this doesn't actually check that it resumes with the right values, as
+        # that's non-trivial... so we just make sure it does *something* when
+        # None is given and that it records whatever it does
+        s.run_mcmc(None, N=self.N)
+        assert s.chain.shape[1] == 2 * self.N