+++ emcee/utils.py
@@ -26,6 +26,23 @@ def sample_ball(p0, std, size=1):
                       for i in range(size)])
 
 
+def sample_ellipsoid(p0, covmat, size=1):
+    """
+    Produce an ellipsoid of walkers around an initial parameter value,
+    according to a covariance matrix.
+
+    :param p0: The initial parameter value.
+    :param covmat:
+        The covariance matrix.  Must be symmetric-positive definite or
+        it will raise the exception numpy.linalg.LinAlgError
+    :param size: The number of samples to produce.
+
+    """
+    return np.random.multivariate_normal(np.atleast_1d(p0),
+                                         np.atleast_2d(covmat),
+                                         size=size)
+
+
 class MH_proposal_axisaligned(object):
     """
     A Metropolis-Hastings proposal, with axis-aligned Gaussian steps,