+++ HISTORY.rst
@@ -1,12 +1,21 @@
 .. :changelog:
 
-2.1.0 (upcoming)
-++++++++++++++++
+2.2.0 (2016-07-12)
+++++++++++++++++++
+
+- Improved autocorrelation time computation.
+- Numpy compatibility issues.
+- Fixed deprecated integer division behavior in PTSampler.
+
+
+2.1.0 (2014-05-22)
+++++++++++++++++++
 
 - Removing dependence on ``acor`` extension.
 - Added arguments to ``PTSampler`` function.
 - Added automatic load-balancing for MPI runs.
 - Added custom load-balancing for MPI and multiprocessing.
+- New default multiprocessing pool that supports ``^C``.
 
 
 2.0.0 (2013-11-17)